Add formula-based stat evaluation system with new `IFormula` implementations and editor support

+++ Reactive Programming/Assets/Scripts/Services/StatResolver.cs
@@ -9,14 +9,14 @@ namespace Economy {
     public class StatResolver {
         public ComputedStats Resolve(BuildingState building, List<StatModifier> modifiers) {
             var result = new ComputedStats {
-                Income = ResolveOne(building.GetLevelBasedValue(StatType.Income), StatType.Income, modifiers),
-                Frequency = ResolveOne(building.GetLevelBasedValue(StatType.Frequency), StatType.Frequency, modifiers),
-                Cost = ResolveOne(building.GetLevelBasedValue(StatType.Cost), StatType.Cost, modifiers),
-                StabilityModifier = ResolveOne(building.GetLevelBasedValue(StatType.StabilityModifier), StatType.StabilityModifier, modifiers),
-                StabilityModifierMultiplier = ResolveOne(building.GetLevelBasedValue(StatType.StabilityModifierMultiplier), StatType.StabilityModifierMultiplier, modifiers),
-                MultiplierCoefficient = ResolveOne(building.GetLevelBasedValue(StatType.MultiplierCoefficient), StatType.MultiplierCoefficient, modifiers),
-                CriticalChance = ResolveOne(building.GetLevelBasedValue(StatType.CriticalChance), StatType.CriticalChance, modifiers),
-                CriticalMultiplier = ResolveOne(building.GetLevelBasedValue(StatType.CriticalMultiplier), StatType.CriticalMultiplier, modifiers)
+                Income = ResolveOne((float)building.GetLevelBasedValue(StatType.Income), StatType.Income, modifiers),
+                Frequency = ResolveOne((float)building.GetLevelBasedValue(StatType.Frequency), StatType.Frequency, modifiers),
+                Cost = ResolveOne((float)building.GetLevelBasedValue(StatType.Cost), StatType.Cost, modifiers),
+                StabilityModifier = ResolveOne((float)building.GetLevelBasedValue(StatType.StabilityModifier), StatType.StabilityModifier, modifiers),
+                StabilityModifierMultiplier = ResolveOne((float)building.GetLevelBasedValue(StatType.StabilityModifierMultiplier), StatType.StabilityModifierMultiplier, modifiers),
+                MultiplierCoefficient = ResolveOne((float)building.GetLevelBasedValue(StatType.MultiplierCoefficient), StatType.MultiplierCoefficient, modifiers),
+                CriticalChance = ResolveOne((float)building.GetLevelBasedValue(StatType.CriticalChance), StatType.CriticalChance, modifiers),
+                CriticalMultiplier = ResolveOne((float)building.GetLevelBasedValue(StatType.CriticalMultiplier), StatType.CriticalMultiplier, modifiers)
             };
 
             result.Frequency = Mathf.Max(0.01f, result.Frequency);
